Getting somewhere with the tether.
Pretty much everything about the tether is functional except I'm still trying to figure out exactly how I want the joints to move. (Which is the biggest thing)

Tried so many ways and the current one was unfortunately the most promising but it still fails badly with basic obstacles and looks strange when moving fast.

• Inactive joints are hidden in the hierarchy

• Better logic when activating and deactivating joints and also limited the amount of times this can happen per frame.

• Added a physics material to the joints (slips around obstacles slightly better)